+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.0.2 / 2012-08-14
+------------------
+* Added `error` method.
+* Made `series()` an alias for `sequential()`.
+
 0.0.1 / 2012-08-14
 ------------------
 * Initial release.